requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorlog==6.9.0
coverage==7.6.7
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
flake8==7.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
identify==2.6.3
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
kombu==5.4.2
Lightflow==1.11.1
liteflow.core==0.3.1
liteflow.providers.mongo==0.3.1
loguru==0.7.2
Mako==1.3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.13.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.0
pymongo==4.10.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-order==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-interface==1.6.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
rich==13.9.4
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
uvloop==0.21.0
vine==5.1.0
virtualenv==20.28.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
